# glTF export performance improvements