--- a/src/main/kotlin/models/glTF/Buffer.kt
+++ b/src/main/kotlin/models/glTF/Buffer.kt
@@ -5,13 +5,26 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 class Buffer(filename: String) {
     val uri = "$filename.bin"
 
-    fun getByteLength() = bytes.size
+    fun getByteLength() = byteChunks.sumOf { it.size }
 
-    // TODO: replace this with something more efficient
-    var bytes = ByteArray(0)
-        @JsonIgnore get
+    private val byteChunks = ArrayList<ByteArray>()
+
+    @JsonIgnore
+    fun getBytes(): ByteArray {
+        val finalBytes = ByteArray(getByteLength())
+        byteChunks.fold(0) { pos, chunk ->
+            System.arraycopy(chunk, 0, finalBytes, pos, chunk.size)
+            pos + chunk.size
+        }
+
+        // Might as well cache the fruit of our efforts
+        byteChunks.clear()
+        byteChunks.add(finalBytes)
+
+        return finalBytes
+    }
 
     fun addBytes(bytesToAdd: ByteArray) {
-        bytes += bytesToAdd
+        byteChunks.add(bytesToAdd)
     }
 }