I timed the export process

Before any changes, exporting 12850 at radius 8:
```
Took 29550330805 nanoseconds
Took 25447103456 nanoseconds
Took 20830986557 nanoseconds
Took 22491631446 nanoseconds
Took 20685962938 nanoseconds
```

Median 22.5 seconds / Mean 23.8 seconds

After first commit:
```
Took 16850471357 nanoseconds
Took 14769895850 nanoseconds
Took 13297833444 nanoseconds
Took 13833327370 nanoseconds
Took 13495152207 nanoseconds
```

Median 13.8 seconds / Mean 14.4 seconds

After second commit:
```
Took 3667631632 nanoseconds
Took 2749942353 nanoseconds
Took 3241219574 nanoseconds
Took 3616193326 nanoseconds
Took 2823936507 nanoseconds
```

Median 3.24 seconds / Mean 3.22 seconds

The above timings were taken with a fairly naïve approach of just calling `System.nanoTime()` like so:

```kt
    fun exportScene() {
        val nanos = System.nanoTime()
        SceneExporter().exportSceneToFile(scene, this)
        val timeDelta = System.nanoTime() - nanos
        println("Took $timeDelta nanoseconds")
    }
```

This means that they include the timings for the full export process from start to finish, including things like copying textures.

---

So there's roughly an 86% speedup in exporting that particular region. I checked the bin and gltf files, and both are still byte-for-byte identical to the ones generated without these commits.